Quantization/requirements.txt
numpy
torch
torchvision
onnxruntime
librosa
matplotlib
Pillow
